backend/requirements.txt
# FastAPI and web framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# Google Gemini AI
google-generativeai==0.7.2

# Supabase
supabase==2.9.1

# Utilities
python-dotenv==1.0.1